requirements.txt
PySide6>=6.5.0
PyOpenGL>=3.1.0
numpy>=1.24.0
